+++ core/changelog.md
@@ -1,3 +1,6 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
+- feat: add numberOfRetries, fallbackIndex and selected key name to context
+[BREAKING] changed BifrostContextKeySelectedKey to BifrostContextKeySelectedKeyID
+- feat: send model deployment back in response extra fields