+++ transports/changelog.md
@@ -1,4 +1,8 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- change: health endpoint is whitelisted from auth middleware
+- chore: version update framework to 1.1.24
+- chore: allowed changing name when updating a virtual key
+- feat: add numberOfRetries, fallbackIndex and selected key name and id to context to telemetry metrics
+- feat: add used virtual key name and id to telemetry metrics
+- feat: send model deployment back in response extra fields